click image working
