+++ app/(auth)/(tabs)/home.tsx
@@ -5,7 +5,8 @@ import { hp, wp } from "@/lib/common";
 import { useRouter } from "expo-router";
 import React from "react";
 import { StyleSheet, Text, View } from "react-native";
-import Recommendation from "../../components/Recommendation";
+import Recommendation from "../../../components/Recommendation";
+import Button from "@/components/Button";
 
 const Home = () => {
   const router = useRouter();
