+++ app/(auth)/viewpost/[postId].tsx
@@ -1,0 +1,9 @@
+import { useLocalSearchParams } from "expo-router";
+import { Text } from "react-native";
+
+const ViewPost = () => {
+  const { postId } = useLocalSearchParams();
+  return <Text>Post ID: {postId}</Text>;
+};
+
+export default ViewPost;
